# swagger.json
{
  "swagger": "2.0",
  "basePath": "/",
...
  "info": {
    "title": "skinswipe",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/getIntersections": {
      "post": {
        "description": "Получение списка людей для ленты",
        "tags": ["people"],
        "consumes": ["application/json"],
        "parameters": [
          {
            "in": "header",
            "name": "x-access-token",
            "type": "string",
            "required": true
          },
          {
            "in": "header",
            "name": "appVersion",
            "type": "string",
            "required": true
          },
          {
            "in": "body",
            "name": "body",
            "description": "Request body.",
            "schema": {
              "type": "object",
              "properties": {
                "offset": {
                  "type": "number"
                },
                "limit": {
                  "type": "number"
                },
                "filters": {
                  "type": "string"
                },
                "username": {
                  "type": "string"
                },
                "online": {
                  "type": "number"
                }
              },
              "required": ["offset", "limit"]
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        }
      }
    },
    "/api/getIntersection": {
      "post": {
        "description": "Получение 1 карточки с конкретным юзером",
        "tags": ["people"],
        "consumes": ["application/json"],
        "parameters": [
          {
            "in": "header",
            "name": "x-access-token",
            "type": "string",
            "required": true
          },
          {
            "in": "header",
            "name": "appVersion",
            "type": "string",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        }
      }
    },
...
    "/api/v2/getIntersections": {
      "post": {
        "description": "Получение списка людей для ленты (-A)",
        "tags": ["people"],
        "consumes": ["application/json"],
        "parameters": [
          {
            "in": "header",
            "name": "x-access-token",
            "type": "string",
            "required": true
          },
          {
            "in": "header",
            "name": "appVersion",
            "type": "string",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        }
      }
    },
...
    "/api/user/subAndroidTest": {
      "post": {
        "description": "Подписка на андроид, устаревший метод",
        "tags": ["subscription"],
        "consumes": ["application/json"],
        "parameters": [
          {
            "in": "header",
            "name": "x-access-token",
            "type": "string",
            "required": true
          },
          {
            "in": "header",
            "name": "appVersion",
            "type": "string",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        }
      }
    },
...
    "/api/user/sub/android/amazon": {
      "post": {
        "description": "Подписка на андроид amazon, новый метод",
        "tags": ["subscription"],
        "consumes": ["application/json"],
        "parameters": [
          {
            "in": "header",
            "name": "x-access-token",
            "type": "string",
            "required": true
          },
          {
            "in": "header",
            "name": "appVersion",
            "type": "string",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        }
      }
    },
    "/api/user/sub/android/amazon/reset": {
      "post": {
        "description": "Удаление amazon подписки",
        "tags": ["subscription"],
        "consumes": ["application/json"],
        "parameters": [
          {
            "in": "header",
            "name": "x-access-token",
            "type": "string",
            "required": true
          },
          {
            "in": "header",
            "name": "appVersion",
            "type": "string",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        }
      }
    },
    "/api/user/subIOS": {
      "post": {
        "description": "Подписка для ios, проверяет в apple",
        "tags": ["subscription"],
        "consumes": ["application/jsonandroid/amazon"],
        "parameters": [
          {
            "in": "header",
            "name": "x-access-token",
            "type": "string",
            "required": true
          },
          {
            "in": "header",
            "name": "appVersion",
            "type": "string",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        }
      }
    },
    "/api/subscription/iosKassa": {
      "post": {
        "description": "Подписка для ios, проверяет в yandex.kassa",
        "tags": ["subscription"],
        "consumes": ["application/json"],
        "parameters": [
          {
            "in": "header",
            "name": "x-access-token",
            "type": "string",
            "required": true
          },
          {
            "in": "header",
            "name": "appVersion",
            "type": "string",
            "required": true
          },
          {
            "in": "body",
            "name": "body",
            "description": "Request body.",
            "schema": {
              "type": "object",
              "properties": {
                "token": {
                  "type": "string"
                }
              },
              "required": ["token"]
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        }
      }
    },
    "/api/subscription/iosKassaToken": {
      "post": {
        "description": "Подписка для ios, записывает токен и выдаёт confirmation_url",
        "tags": ["subscription"],
        "consumes": ["application/json"],
        "parameters": [
          {
            "in": "header",
            "name": "x-access-token",
            "type": "string",
            "required": true
          },
          {
            "in": "header",
            "name": "appVersion",
            "type": "string",
            "required": true
          },
          {
            "in": "body",
            "name": "body",
            "description": "Request body.",
            "schema": {
              "type": "object",
              "properties": {
                "token": {
                  "type": "string"
                },
                "amount": {
                  "type": "string"
                },
                "months": {
                  "type": "number"
                }
              },
              "required": ["token", "amount", "months"]
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        }
      }
    },
    "/api/user/getSubState": {
      "post": {
        "description": "Получение информации о подписке, используется при старте приложения",
        "tags": ["subscription"],
        "consumes": ["application/json"],
        "parameters": [
          {
            "in": "header",
            "name": "x-access-token",
            "type": "string",
            "required": true
          },
          {
            "in": "header",
            "name": "appVersion",
            "type": "string",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        }
      }
    },
...
  }
}